feat(mobile): :sparkles: put share/report on bottom #1493
Fixes #1493
