+++ mobile/routes/text.js
@@ -1,8 +1,6 @@
-import { ENDPOINT, HOST } from '../constants/urls.js'
 import { Button } from 'react-native-paper'
 import React, { Fragment } from 'react'
 import SCREENS from '../constants/screens.js'
-import { Share } from 'react-native'
 import TextDrawer from './text-drawer.js'
 import UI from '../constants/ui.js'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
@@ -14,18 +12,6 @@ const selector = (state) => state.text
 
 export default function Text() {
   const { text } = useSelector(selector)
-  const shareLabel = 'Share'
-
-  const onShare = async () => {
-    try {
-      await Share.share({
-        message: text.title,
-        url: `${HOST.frontend}/${ENDPOINT.texts}/${text.slug}`
-      })
-    } catch (error) {
-      alert(error.message)
-    }
-  }
 
   return (
     <Stack.Navigator initialRouteName={SCREENS.home}>
@@ -42,7 +28,7 @@ export default function Text() {
 
           headerRight: () => (
             <Fragment>
-              <Button onPress={onShare}>{shareLabel}</Button>
+              <Button icon="cog" />
             </Fragment>
           )
         }}
